working db connection
