+++ docker-compose-dev-m1.yaml
@@ -44,7 +44,7 @@ services:
     build:
       context: server/docs
     ports:
-      - '8081:8081'
+      - '8082:8082'
     labels:
       - "traefik.http.routers.server-docs.rule=Host(`server-docs.docker.localhost`)"
   swagger-ui:
@@ -69,15 +69,15 @@ services:
         read_only: true
     labels:
       - "traefik.http.routers.prometheus.rule=Host(`prometheus.docker.localhost`)"
-  go-prometheus:
-    build:
-      context: server/prometheus-microservice
-    ports:
-      - "9091:9091"
-    volumes:
-      - ./server/prometheus-microservice:/server/prometheus
-    labels:
-      - "traefik.http.routers.go-prometheus.rule=Host(`go-prometheus.docker.localhost`)"
+#  go-prometheus:
+#    build:
+#      context: server/prometheus-microservice
+#    ports:
+#      - "9091:9091"
+#    volumes:
+#      - ./server/prometheus-microservice:/server/prometheus
+#    labels:
+#      - "traefik.http.routers.go-prometheus.rule=Host(`go-prometheus.docker.localhost`)"
   reverse-proxy:
     image: traefik:v2.4
     command:
