+++ server/run-dev.sh
@@ -1,7 +1,6 @@
 #!/bin/sh
 
 sh -c ./bin/users
-sh -c ./bin/prometheus
 
 # Along with creating new microservices,
 # simply add another file to be executed
