+++ server/users/conf/conf_dev.ini
@@ -5,4 +5,4 @@ WriteTimeout = 60
 MaxHeaderBytes = 1 << 20
 
 [mongodb]
-Uri = mongodb://localhost:27017
+Uri = mongodb://root:example@mongodb:27017
